BackEnd-Flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
bcrypt==4.0.1
marshmallow==3.20.1
